etl/pyproject.toml
[project]
name = "ad-manager-etl"
version = "0.1.0"
description = "広告データETLパイプライン（Google Ads / Yahoo!広告 / Bing Ads → BigQuery）"
requires-python = ">=3.12"
dependencies = [
    "google-ads>=25.0.0",
    "google-cloud-bigquery>=3.20.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-logging>=3.10.0",
    "bingads>=13.0.18",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
